refactor: rename files, rm forward definition
